requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
aiogram==3.13.1
pydantic==2.9.2
python-dotenv==1.0.1
aiomysql==0.2.0
loguru==0.7.2